Handle Error related to using the "demo" API Key (#2)
**Revision Summary:**

We noticed that the program crashes when it uses "demo" as the API Key. From my memory, it had previously been OK to use the "demo" API Key programmatically and the API wouldn't restrict access, but perhaps the AlphaVantage API changed its behavior. Now the normal browser message is returned:

```py
{'Information': 'The **demo** API key is for demo purposes only. Please claim your free API key at (https://www.alphavantage.co/support/#api-key) to explore our full API offerings. It takes fewer than 20 seconds, and we are committed to making it free forever.'}
```

So the program either needs to not default to using the "demo" API Key, or the program should gracefully handle errors resulting from using this key.

The way to handle the error would be something like the following:

```py
    if "Information" in parsed_response.keys() and "Please claim your free API key" in parsed_response["Information"]:
        # HANDLE A RESPONSE ERROR DUE TO MISSING API KEY
        print(parsed_response["Information"])
        exit()
```

But an easier and simpler solution is to just default to not specifying an API Key. In theory this would result in a runtime error as well, but the `None` value is actually converted to a string `"None"` when concatenated into the request URL, and this is surprisingly a valid API Key.

I guess another, better solution would be to use "abc123" (another surprisingly valid API Key) as the default value.

**Commit History:**

* The demo default is actually more buggy than a blank env var, due to the way the API is configured to accept a wide range of API keys, one of which is a string version of None, but using the demo key actually triggers an error
* Improve resiliency